Split out functions from graph_task.py into two files, create_tasks.py and execute_tasks.py for org purposes, but primarily at this stage to make sure the TraversalNode and the ExecutionNodes are being used in the proper places.
