+++ src/fides/api/task/graph_task.py
@@ -1,17 +1,12 @@
 # pylint: disable=too-many-lines
 import copy
-import json
 import traceback
 from abc import ABC
 from functools import wraps
 from time import sleep
 from typing import Any, Callable, Dict, List, Optional, Set, Tuple, Union
 
-import dask
-import networkx
-from dask import delayed  # type: ignore[attr-defined]
 from loguru import logger
-from networkx import NetworkXNoCycle
 from sqlalchemy.orm import Query, Session
 
 from fides.api.common_exceptions import (
@@ -42,11 +37,10 @@ from fides.api.models.connectionconfig import (
     ConnectionType,
 )
 from fides.api.models.datasetconfig import DatasetConfig
-from fides.api.models.policy import CurrentStep, Policy
+from fides.api.models.policy import Policy
 from fides.api.models.privacy_request import (
     ExecutionLogStatus,
     PrivacyRequest,
-    PrivacyRequestStatus,
     RequestTask,
     TaskStatus,
 )
@@ -70,8 +64,6 @@ from fides.api.util.logger import Pii
 from fides.api.util.saas_util import FIDESOPS_GROUPED_INPUTS
 from fides.api.util.storage_util import storage_json_encoder
 from fides.config import CONFIG
-
-dask.config.set(scheduler="threads")
 
 COLLECTION_FIELD_PATH_MAP = Dict[CollectionAddress, List[Tuple[FieldPath, FieldPath]]]
 
@@ -727,467 +719,6 @@ def start_function(seed: List[Dict[str, Any]]) -> Callable[[], List[Dict[str, An
     return g
 
 
-def collect_tasks_fn(
-    tn: TraversalNode, data: Dict[CollectionAddress, TraversalNode]
-) -> None:
-    """Run the traversal, as an action returning the modified traversal node"""
-    if not tn.is_root_node():
-        data[tn.address] = tn
-
-
-def run_access_request(
-    privacy_request: PrivacyRequest,
-    policy: Policy,
-    graph: DatasetGraph,
-    connection_configs: List[ConnectionConfig],
-    identity: Dict[str, Any],
-    session: Session,
-) -> None:
-    """Queue the access request nodes"""
-    traversal: Traversal = Traversal(graph, identity)
-
-    env: Dict[CollectionAddress, TraversalNode] = {}
-    # Traversal.traverse modifies "env" in place, adding parents and children to the traversal nodes
-    end_nodes: List[CollectionAddress] = traversal.traverse(env, collect_tasks_fn)
-
-    created_privacy_request_tasks = create_access_request_task_objects(
-        session, privacy_request, traversal, env, end_nodes, graph
-    )
-    for task in created_privacy_request_tasks:
-        if task.upstream_tasks_complete(session):
-            logger.info(
-                f"Queuing access task {privacy_request.id} > {task.id}: {task.collection_address}"
-            )
-            run_access_node.delay(privacy_request.id, task.id)
-    return
-
-
-def build_access_networkx_digraph(
-    env: Dict[CollectionAddress, TraversalNode],
-    end_nodes: List[CollectionAddress],
-    traversal: Traversal,
-) -> networkx.DiGraph:
-    """
-    Builds a networkx graph to get consistent formatting of upstream/downstream nodes, regardless of whether the node
-    is an actual node, or a "root"/"terminator" node.
-
-    Primarily though, we can utilize networkx's "descendants" method to pre-calculate every downstream node that
-    can be reached by each current node
-    """
-    first_nodes: Dict[FieldAddress, str] = traversal.extract_seed_field_addresses()
-
-    networkx_graph = networkx.DiGraph()
-    networkx_graph.add_nodes_from(env.keys())
-    networkx_graph.add_nodes_from([TERMINATOR_ADDRESS, ROOT_COLLECTION_ADDRESS])
-
-    for node in [
-        CollectionAddress(initial_node.dataset, initial_node.collection)
-        for initial_node in first_nodes
-    ]:
-        networkx_graph.add_edge(ROOT_COLLECTION_ADDRESS, node)
-
-    for collection_address, traversal_node in env.items():
-        for child in traversal_node.children:
-            networkx_graph.add_edge(collection_address, child)
-
-    for node in end_nodes:
-        networkx_graph.add_edge(node, TERMINATOR_ADDRESS)
-
-    return networkx_graph
-
-
-def create_access_request_task_objects(
-    session: Session,
-    privacy_request: PrivacyRequest,
-    traversal: Traversal,
-    env: Dict[CollectionAddress, Any],
-    end_nodes: List[CollectionAddress],
-    dataset_graph: DatasetGraph,
-):
-    """Add PrivacyRequestTasks to the database or update existing Privacy Request tasks to a pending status
-
-    There should only be one completed task for each node of each action type.
-
-    This should only be called if a PrivacyRequest is in a pending or errored state.
-    """
-    graph: networkx.DiGraph = build_access_networkx_digraph(env, end_nodes, traversal)
-
-    ready_tasks: List[RequestTask] = []
-
-    for node in list(networkx.topological_sort(graph)):
-        existing_completed_task = (
-            session.query(RequestTask)
-            .filter(
-                RequestTask.privacy_request_id == privacy_request.id,
-                RequestTask.action_type == ActionType.access,
-                RequestTask.collection_address == node.value,
-                RequestTask.status == TaskStatus.complete,
-            )
-            .first()
-        )
-
-        if existing_completed_task:
-            continue
-
-        json_collection: Optional[Dict] = None
-        if node not in [ROOT_COLLECTION_ADDRESS, TERMINATOR_ADDRESS]:
-            json_collection = json.loads(dataset_graph.nodes[node].collection.json())
-
-        task = RequestTask.create(
-            session,
-            data={
-                "privacy_request_id": privacy_request.id,
-                "upstream_tasks": [
-                    upstream.value for upstream in graph.predecessors(node)
-                ],
-                "downstream_tasks": [
-                    downstream.value for downstream in graph.successors(node)
-                ],
-                "dataset_name": node.dataset,
-                "collection_name": node.collection,
-                "status": TaskStatus.complete
-                if node == ROOT_COLLECTION_ADDRESS
-                else TaskStatus.pending,
-                "access_data": [traversal.seed_data]
-                if node == ROOT_COLLECTION_ADDRESS
-                else [],
-                "action_type": ActionType.access,
-                "collection_address": node.value,
-                "all_descendant_tasks": [
-                    descend.value for descend in list(networkx.descendants(graph, node))
-                ],
-                "collection": json_collection,
-                "traversal_details": build_traversal_details(node, env),
-            },
-        )
-        ready_tasks.append(task)
-
-    return ready_tasks
-
-
-def build_traversal_details(
-    node: CollectionAddress, env: Dict[CollectionAddress, TraversalNode]
-):
-    """Pull some key elements off of the traversal node for use in execution on the individual task"""
-    if node in [ROOT_COLLECTION_ADDRESS, TERMINATOR_ADDRESS]:
-        return None
-
-    traversal_node: TraversalNode = env[node]
-    connection_key = traversal_node.node.dataset.connection_key
-
-    return TraversalDetails(
-        dataset_connection_key=connection_key,
-        incoming_edges=[
-            [edge.f1.value, edge.f2.value] for edge in traversal_node.incoming_edges()
-        ],
-        outgoing_edges=[
-            [edge.f1.value, edge.f2.value] for edge in traversal_node.outgoing_edges()
-        ],
-        input_keys=[tn.value for tn in traversal_node.input_keys()],
-    ).dict()
-
-
-def build_erasure_networkx_digraph(
-    env: Dict[CollectionAddress, Any],
-    end_nodes: List[CollectionAddress],
-) -> networkx.DiGraph:
-    """
-    Builds a networkx graph to get consistent formatting of upstream/downstream nodes, regardless of whether the node
-    is an actual node, or a "root"/"terminator" node.
-
-    Primarily though, we can utilize networkx's "descendants" method to pre-calculate every downstream node that
-    can be reached by each current node
-    """
-    networkx_graph = networkx.DiGraph()
-    networkx_graph.add_nodes_from(env.keys())
-    networkx_graph.add_nodes_from([ROOT_COLLECTION_ADDRESS, TERMINATOR_ADDRESS])
-
-    for node_name, traversal_node in env.items():
-        # Modifies end nodes in place as well?
-        erasure_dependencies = _evaluate_erasure_dependencies(traversal_node, end_nodes)
-        for dep in erasure_dependencies:
-            networkx_graph.add_edge(dep, node_name)
-
-    for node in end_nodes:
-        networkx_graph.add_edge(node, TERMINATOR_ADDRESS)
-
-    try:
-        networkx.find_cycle(networkx_graph, ROOT_COLLECTION_ADDRESS)
-    except NetworkXNoCycle:
-        logger.info("No cycles found as expected")
-    else:
-        raise Exception("Cycles found in erasure DAG")
-
-    return networkx_graph
-
-
-def upfront_checks(session, privacy_request_id, privacy_request_task_id):
-    privacy_request: PrivacyRequest = PrivacyRequest.get(
-        db=session, object_id=privacy_request_id
-    )
-    request_task: RequestTask = RequestTask.get(
-        db=session, object_id=privacy_request_task_id
-    )
-
-    if not privacy_request:
-        raise PrivacyRequestNotFound(
-            f"Privacy request with id {privacy_request_id} not found"
-        )
-
-    if not request_task:
-        raise RequestTaskNotFound(f"Request Task with id {request_task} not found")
-
-    upstream_results = session.query(RequestTask).filter(False)
-    if request_task.status == TaskStatus.pending:
-        upstream_results: Query = session.query(RequestTask).filter(
-            RequestTask.privacy_request_id == privacy_request_id,
-            RequestTask.status == PrivacyRequestStatus.complete,
-            RequestTask.action_type == request_task.action_type,
-            RequestTask.collection_address.in_(request_task.upstream_tasks),
-        )
-
-        if not upstream_results.count() == len(request_task.upstream_tasks):
-            raise RequestTaskNotFound(
-                f"Cannot start Privacy Request Task {request_task.id} - status is {request_task.status}"
-            )
-    return privacy_request, request_task, upstream_results
-
-
-def get_task_resources(session: Session, privacy_request: PrivacyRequest):
-    return TaskResources(
-        privacy_request,
-        privacy_request.policy,
-        session.query(ConnectionConfig).all(),
-        session,
-    )
-
-
-@celery_app.task(base=DatabaseTask, bind=True)
-def run_erasure_node(
-    self: DatabaseTask, privacy_request_id: str, privacy_request_task_id: str
-) -> None:
-    with self.get_new_session() as session:
-        privacy_request, request_task, upstream_results = upfront_checks(
-            session, privacy_request_id, privacy_request_task_id
-        )
-        graph_task_key = CollectionAddress.from_string(request_task.collection_address)
-
-        if request_task.status == TaskStatus.complete:
-            logger.info(
-                f"Cannot start erasure Privacy Request Task {request_task.id} {request_task.collection_address} - status is {request_task.status}"
-            )
-            return
-
-        task_resources = get_task_resources(session, privacy_request)
-
-        logger.info(
-            f"Running erasure task {request_task.collection_address}:{request_task.id} for privacy request {privacy_request.id}"
-        )
-
-        if graph_task_key != TERMINATOR_ADDRESS:
-            graph_task: GraphTask = GraphTask(request_task, task_resources)
-            func: Callable = graph_task.erasure_request
-            retrieved_data: List[Row] = request_task.data_for_erasures or []
-            inputs: List[List[Row]] = request_task.erasure_input_data or []
-            func(retrieved_data, inputs)
-            request_task.status = TaskStatus.complete
-            request_task.save(session)
-            logger.info(f"Marking node {graph_task_key} as complete")
-
-        else:
-            request_task.status = TaskStatus.complete
-            request_task.save(session)
-            logger.info(f"Marking node {graph_task_key} as complete")
-
-            from fides.api.service.privacy_request.request_runner_service import (
-                queue_privacy_request,
-            )
-
-            logger.info(
-                f"Terminator node reached for erasure graph of {privacy_request.id}. Queueing."
-            )
-
-            queue_privacy_request(
-                privacy_request_id=privacy_request.id,
-                from_step=CurrentStep.finalize_erasure.value,
-            )
-            return
-
-        # If there are other downstream tasks that are ready to go, queue those now
-        pending_downstream_tasks: Query = request_task.pending_downstream_tasks(session)
-        for downstream_task in pending_downstream_tasks:
-            if downstream_task.upstream_tasks_complete(session):
-                logger.info(f"Upstream tasks complete so queuing {downstream_task.id}")
-                run_erasure_node.delay(
-                    privacy_request_id=privacy_request_id,
-                    privacy_request_task_id=downstream_task.id,
-                )
-            else:
-                logger.info(
-                    f"Upstream tasks not completed for {downstream_task.collection_address} {downstream_task.id}"
-                )
-
-
-@celery_app.task(base=DatabaseTask, bind=True)
-def run_access_node(
-    self: DatabaseTask, privacy_request_id: str, privacy_request_task_id: str
-) -> None:
-    with self.get_new_session() as session:
-        privacy_request, request_task, upstream_results = upfront_checks(
-            session, privacy_request_id, privacy_request_task_id
-        )
-
-        graph_task_key = CollectionAddress.from_string(request_task.collection_address)
-
-        if request_task.status != TaskStatus.pending:
-            logger.info(
-                f"Cannot start access Privacy Request Task {request_task.id} {request_task.collection_address} - status is {request_task.status}"
-            )
-            return
-
-        task_resources = get_task_resources(session, privacy_request)
-
-        logger.info(
-            f"Running access request task {request_task.id}:{request_task.collection_address} for privacy request {privacy_request.id}"
-        )
-
-        if graph_task_key == TERMINATOR_ADDRESS:
-            # For the terminator node, collect all of the results
-            final_results: Dict = {}
-            for task in privacy_request.access_tasks.filter(
-                RequestTask.status == PrivacyRequestStatus.complete,
-                RequestTask.collection_address.notin_(
-                    [ROOT_COLLECTION_ADDRESS.value, TERMINATOR_ADDRESS.value]
-                ),
-            ):
-                final_results[task.collection_address] = task.access_data
-
-            request_task.terminator_data = final_results
-        else:
-            # For regular nodes, store the data on the node
-            graph_task: GraphTask = GraphTask(request_task, task_resources)
-            func = graph_task.access_request
-            ordered_upstream_tasks = order_tasks_by_input_key(
-                graph_task.execution_node.input_keys, upstream_results
-            )
-            # Put access data in the same order as the input keys
-            func(
-                *[
-                    upstream.access_data if upstream else []
-                    for upstream in ordered_upstream_tasks
-                ]
-            )
-
-        request_task.status = TaskStatus.complete
-        request_task.save(session)
-
-        # If there are other downstream tasks that are ready to go, queue those now
-        pending_downstream_tasks: Query = request_task.pending_downstream_tasks(session)
-        for downstream_task in pending_downstream_tasks:
-            if downstream_task.upstream_tasks_complete(session):
-                logger.info(f"Upstream tasks complete so queuing {downstream_task.id}")
-                run_access_node.delay(
-                    privacy_request_id=privacy_request_id,
-                    privacy_request_task_id=downstream_task.id,
-                )
-            else:
-                logger.info(f"Upstream tasks not completed for {downstream_task.id}")
-
-        # If we've made it to the terminator node, requeue the entire Privacy request to continue where we left off,
-        # by uploading the current access results.
-        if graph_task_key == TERMINATOR_ADDRESS:
-            from fides.api.service.privacy_request.request_runner_service import (
-                queue_privacy_request,
-            )
-
-            queue_privacy_request(
-                privacy_request_id=privacy_request.id,
-                from_step=CurrentStep.upload_access.value,
-            )
-
-
-@celery_app.task(base=DatabaseTask, bind=True)
-def run_consent_node(
-    self: DatabaseTask, privacy_request_id: str, privacy_request_task_id: str
-) -> None:
-    with self.get_new_session() as session:
-        privacy_request, request_task, upstream_results = upfront_checks(
-            session, privacy_request_id, privacy_request_task_id
-        )
-
-        graph_task_key = CollectionAddress.from_string(request_task.collection_address)
-
-        if request_task.status != TaskStatus.pending:
-            logger.info(
-                f"Cannot start consent Privacy Request Task {request_task.id} {request_task.collection_address} - status is {request_task.status}"
-            )
-            return
-
-        logger.info(
-            f"Running consent request task {request_task.id}:{request_task.collection_address} for privacy request {privacy_request.id}"
-        )
-
-        if graph_task_key != TERMINATOR_ADDRESS:
-            task_resources = TaskResources(
-                privacy_request,
-                privacy_request.policy,
-                session.query(ConnectionConfig).all(),
-                session,
-            )
-            task = GraphTask(request_task, task_resources)
-
-            # For regular nodes, store the data on the node
-            func = task.consent_request
-            task.request_task = request_task
-
-            # TODO catch errors if no upstream result
-            func(upstream_results[0].consent_data)
-
-        request_task.status = TaskStatus.complete
-        request_task.save(session)
-
-        # If there are other downstream tasks that are ready to go, queue those now
-        pending_downstream_tasks: Query = request_task.pending_downstream_tasks(session)
-        for downstream_task in pending_downstream_tasks:
-            if downstream_task.upstream_tasks_complete(session):
-                logger.info(f"Upstream tasks complete so queuing {downstream_task.id}")
-                run_consent_node.delay(
-                    privacy_request_id=privacy_request_id,
-                    privacy_request_task_id=downstream_task.id,
-                )
-            else:
-                logger.info(f"Upstream tasks not completed for {downstream_task.id}")
-
-        # If we've made it to the terminator node, requeue the entire Privacy request to continue where we left off,
-        if graph_task_key == TERMINATOR_ADDRESS:
-            from fides.api.service.privacy_request.request_runner_service import (
-                queue_privacy_request,
-            )
-
-            queue_privacy_request(
-                privacy_request_id=privacy_request.id,
-                from_step=CurrentStep.finalize_consent.value,
-            )
-
-
-def order_tasks_by_input_key(
-    input_keys: List[CollectionAddress], upstream_tasks: Query
-) -> List[Optional[RequestTask]]:
-    """Order tasks by input key. If task doesn't exist, add None in its place"""
-    tasks: List[Optional[RequestTask]] = []
-    for key in input_keys:
-        task = next(
-            (
-                upstream
-                for upstream in upstream_tasks
-                if upstream.collection_address == key.value
-            ),
-            None,
-        )
-        tasks.append(task)
-    return tasks
-
-
 def filter_by_enabled_actions(
     access_results: Dict[str, Any], connection_configs: List[ConnectionConfig]
 ) -> Dict[str, Any]:
@@ -1243,271 +774,6 @@ def update_erasure_mapping_from_cache(
         dsk[CollectionAddress.from_string(collection_name)] = cached_erasures[
             collection_name
         ]
-
-
-def create_erasure_request_task_objects(
-    session: Session,
-    privacy_request: PrivacyRequest,
-    env: Dict[CollectionAddress, TraversalNode],
-    end_nodes: List[CollectionAddress],
-    dataset_graph: DatasetGraph,
-):
-    """
-    Erasure tasks need to access the Access Tasks.erasure_data
-
-
-    """
-    graph: networkx.DiGraph = build_erasure_networkx_digraph(env, end_nodes)
-
-    ready_tasks: List[RequestTask] = []
-
-    for node in list(networkx.topological_sort(graph)):
-        existing_completed_task = (
-            session.query(RequestTask)
-            .filter(
-                RequestTask.privacy_request_id == privacy_request.id,
-                RequestTask.action_type == ActionType.erasure,
-                RequestTask.collection_address == node.value,
-                RequestTask.status == TaskStatus.complete,
-            )
-            .first()
-        )
-
-        if existing_completed_task:
-            continue
-
-        # Select access task of the same name
-        erasure_task = (
-            session.query(RequestTask)
-            .filter(
-                RequestTask.privacy_request_id == privacy_request.id,
-                RequestTask.action_type == ActionType.access,
-                RequestTask.collection_address == node.value,
-                RequestTask.status == TaskStatus.complete,
-            )
-            .first()
-        )
-        retrieved_task_data = []
-        if erasure_task:
-            retrieved_task_data = erasure_task.data_for_erasures
-
-        # Select upstream inputs of this node for things like email connectors
-        input_tasks = session.query(RequestTask).filter(
-            RequestTask.privacy_request_id == privacy_request.id,
-            RequestTask.action_type == ActionType.access,
-            RequestTask.collection_address.in_(erasure_task.upstream_tasks),
-            RequestTask.status == TaskStatus.complete,
-        )
-        ordered_upstream_tasks = (
-            []
-            if node in [ROOT_COLLECTION_ADDRESS, TERMINATOR_ADDRESS]
-            else order_tasks_by_input_key(env[node].input_keys(), input_tasks)
-        )
-
-        combined_input_data = []
-        for input_data in ordered_upstream_tasks or []:
-            combined_input_data.append(input_data.data_for_erasures or [])
-
-        json_collection: Optional[Dict] = None
-        if node not in [ROOT_COLLECTION_ADDRESS, TERMINATOR_ADDRESS]:
-            json_collection = json.loads(dataset_graph.nodes[node].collection.json())
-
-        task = RequestTask.create(
-            session,
-            data={
-                "privacy_request_id": privacy_request.id,
-                "upstream_tasks": [
-                    upstream.value for upstream in graph.predecessors(node)
-                ],
-                "downstream_tasks": [
-                    downstream.value for downstream in graph.successors(node)
-                ],
-                "dataset_name": node.dataset,
-                "collection_name": node.collection,
-                "status": TaskStatus.complete
-                if node == ROOT_COLLECTION_ADDRESS
-                else TaskStatus.pending,
-                "data_for_erasures": retrieved_task_data or [],
-                "erasure_input_data": combined_input_data,
-                "action_type": ActionType.erasure,
-                "collection_address": node.value,
-                "all_descendant_tasks": [
-                    descend.value for descend in list(networkx.descendants(graph, node))
-                ],
-                "collection": json_collection,
-                "traversal_details": build_traversal_details(node, env),
-            },
-        )
-        ready_tasks.append(task)
-
-    return ready_tasks
-
-
-def create_consent_request_task_objects(
-    session: Session,
-    privacy_request: PrivacyRequest,
-    env: Dict[CollectionAddress, TraversalNode],
-    identity: Dict[str, Any],
-    dataset_graph: DatasetGraph,
-):
-    """
-    The graph built is very simple: there are no relationships between the nodes, every node has
-    identity data input and every node outputs whether the consent request succeeded.
-
-    The DatasetGraph passed in is expected to have one Node per Dataset.  That Node is expected to carry out requests
-    for the Dataset as a whole.
-
-    """
-    ready_tasks: List[RequestTask] = []
-    node_keys: List[CollectionAddress] = [node for node in list(env.keys())]
-    for node in node_keys + [ROOT_COLLECTION_ADDRESS, TERMINATOR_ADDRESS]:
-        existing_completed_task = (
-            session.query(RequestTask)
-            .filter(
-                RequestTask.privacy_request_id == privacy_request.id,
-                RequestTask.action_type == ActionType.consent,
-                RequestTask.collection_address == node.value,
-                RequestTask.status == TaskStatus.complete,
-            )
-            .first()
-        )
-
-        if existing_completed_task:
-            continue
-
-        json_collection: Optional[Dict] = None
-        if node == ROOT_COLLECTION_ADDRESS:
-            upstream_tasks = []
-            downstream_tasks = [node.value for node in node_keys]
-        elif node == TERMINATOR_ADDRESS:
-            upstream_tasks = [node.value for node in node_keys]
-            downstream_tasks = []
-        else:
-            upstream_tasks = [ROOT_COLLECTION_ADDRESS.value]
-            downstream_tasks = [TERMINATOR_ADDRESS.value]
-            if node not in [ROOT_COLLECTION_ADDRESS, TERMINATOR_ADDRESS]:
-                json_collection = json.loads(
-                    dataset_graph.nodes[node].collection.json()
-                )
-
-        task = RequestTask.create(
-            session,
-            data={
-                "privacy_request_id": privacy_request.id,
-                "upstream_tasks": upstream_tasks,
-                "downstream_tasks": downstream_tasks,
-                "dataset_name": node.dataset,
-                "collection_name": node.collection,
-                "status": TaskStatus.complete
-                if node == ROOT_COLLECTION_ADDRESS
-                else TaskStatus.pending,
-                "consent_data": identity if node == ROOT_COLLECTION_ADDRESS else {},
-                "action_type": ActionType.consent,
-                "collection_address": node.value,
-                "all_descendant_tasks": [TERMINATOR_ADDRESS.value],
-                "collection": json_collection,
-                "traversal_details": build_traversal_details(node, env),
-            },
-        )
-        ready_tasks.append(task)
-
-    return ready_tasks
-
-
-def run_erasure(  # pylint: disable = too-many-arguments
-    privacy_request: PrivacyRequest,
-    policy: Policy,
-    graph: DatasetGraph,
-    connection_configs: List[ConnectionConfig],
-    identity: Dict[str, Any],
-    access_request_data: Dict[str, List[Row]],
-    session: Session,
-) -> List[RequestTask]:
-    """Queue the erasure request nodes"""
-    logger.info("Building erasure graph")
-    traversal: Traversal = Traversal(graph, identity)
-
-    env: Dict[CollectionAddress, TraversalNode] = {}
-    # Run traversal.traverse which modifies env in place to have all the nodes linked in the proper order
-    traversal.traverse(env, collect_tasks_fn)
-    erasure_end_nodes = list(graph.nodes.keys())
-
-    created_privacy_request_tasks = create_erasure_request_task_objects(
-        session, privacy_request, env, erasure_end_nodes, graph
-    )
-
-    queued_tasks = []
-    for task in created_privacy_request_tasks:
-        if (
-            task.upstream_tasks_complete(session)
-            and task.collection_address != ROOT_COLLECTION_ADDRESS.value
-        ):
-            logger.info(
-                f"Queuing erasure task {privacy_request.id} > {task.id}: {task.collection_address}"
-            )
-            queued_tasks.append(task)
-            run_erasure_node.delay(privacy_request.id, task.id)
-
-    return queued_tasks
-
-
-def _evaluate_erasure_dependencies(
-    traversal_node: TraversalNode, end_nodes: List[CollectionAddress]
-) -> Set[CollectionAddress]:
-    """
-    Return a set of collection addresses corresponding to collections that need
-    to be erased before the given task. Remove the dependent collection addresses
-    from `end_nodes` so they can be executed in the correct order. If a task does
-    not have any dependencies it is linked directly to the root node
-    """
-    erase_after = traversal_node.node.collection.erase_after
-    for collection in erase_after:
-        if collection in end_nodes:
-            # end_node list is modified in place
-            end_nodes.remove(collection)
-    # this task will execute after the collections in `erase_after` or
-    # execute at the beginning by linking it to the root node
-    return erase_after if len(erase_after) else {ROOT_COLLECTION_ADDRESS}
-
-
-def run_consent_request(  # pylint: disable = too-many-arguments
-    privacy_request: PrivacyRequest,
-    policy: Policy,
-    graph: DatasetGraph,
-    connection_configs: List[ConnectionConfig],
-    identity: Dict[str, Any],
-    session: Session,
-) -> List[RequestTask]:
-    """Run a consent request
-
-    The graph built is very simple: there are no relationships between the nodes, every node has
-    identity data input and every node outputs whether the consent request succeeded.
-
-    The DatasetGraph passed in is expected to have one Node per Dataset.  That Node is expected to carry out requests
-    for the Dataset as a whole.
-    """
-    env = {}
-    for col_address, node in graph.nodes.items():
-        traversal_node = TraversalNode(node)
-        env[col_address] = traversal_node
-
-    created_consent_tasks = create_consent_request_task_objects(
-        session, privacy_request, env, identity, graph
-    )
-
-    queued_tasks = []
-    for task in created_consent_tasks:
-        if (
-            task.upstream_tasks_complete(session)
-            and task.collection_address != ROOT_COLLECTION_ADDRESS.value
-        ):
-            logger.info(
-                f"Queuing consent task {privacy_request.id} > {task.id}: {task.collection_address}"
-            )
-            queued_tasks.append(task)
-            run_consent_node.delay(privacy_request.id, task.id)
-
-    return queued_tasks
 
 
 def build_affected_field_logs(
